requirements.txt
crewai
pandas
openpyxl
schedule
langchain
openai
python-dotenv
langchain-community
smtplib
sendgrid
streamlit